fix(project-agent): split proposal persistence out of the execution coordinator
Merging origin/main pushed projectAgentExecutionCoordinator.ts to 801 lines
(branch 798 + main 789 both under the 800 limit on their own), tripping the
R12 file-size gate on the merge tree.

Root fix is the split R9 asks for, not a raised baseline: persistApprovedProposal
and persistPreparedProposal move to projectAgentProposalPersistence.ts and take
the clock plus the two mutation primitives as an explicit context instead of
closing over the coordinator. Coordinator drops to 710 lines and sheds the
imports the moved code owned.

Also locks the projectAgentState.ts ratchet at its current 808 lines, as the
gate asked.
